backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyMuPDF==1.24.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
requests==2.31.0
pydantic==2.7.1
